\name{getaprior}
\title{IVBART for continuous treatments and outcomes}
\alias{getaprior}

\description{ Returns an \eqn{\alpha}{alpha} concentration parameter prior.
}

\usage{
getaprior(
    n, Imin, Imax, psi, gs
)
}

\arguments{
   \item{n}{ }
   \item{Imin}{ The minimum number of clusters; typically, 1. }
   \item{Imax}{ The maximum number of clusters. }
   \item{psi}{ }
   \item{gs}{ The grid size. }
}

\author{
Robert McCulloch: \email{robert.e.mcculloch@gmail.com},\cr
Rodney Sparapani: \email{rsparapa@mcw.edu}.
}
